Demo of our attack in physical world
The experimenter impersonating Arnold Schwarzenegger. 
The L2-distance computed with FaceNet network between attacker and target are displayed over the image of the target. 
Firstly, The L2-distance is above 1.242 between the original attacker and the target.   
Secondly, when placed the part closest to our patch from the target image on the attaker, the L2-distance is still above 1.242 between the attacker and the target.
Finally, after placed our adversarial patches, the L2-distance is less than 1.242 between the attacker and the target.